Fixed camera aspect ration on resing of window.

+++ PrinciplesOfCG/Camera.cpp
@@ -39,7 +39,7 @@ void Camera::setPerspective(float degrees, float aspectRatio, float zNear, float
 
 void Camera::setDimensions(int width, int heights)
 {
-    this->aspectRatio = (int)width / heights;
+    this->aspectRatio = width / heights;
     forceUpdate();
 }
